Package scripts changes

+++ Scripts/package.ps1
@@ -2,8 +2,20 @@
 # Script parameters
 param(
     [string] $version,
-    [bool]   $publish = $true
+    [switch] $deploy,
+    [switch] $allowDirty
 )
+# ------------------------------------------------------------------------------
+
+
+# ------------------------------------------------------------------------------
+# Exit codes
+
+#   0   OK
+#   1   No Git TAG on repository
+#   2   Invalid version format
+#   3   Repository has pending changes
+
 # ------------------------------------------------------------------------------
 
 
@@ -22,7 +34,7 @@ $project      = "SEnRestAPI"
 $gitHash      = (git rev-parse HEAD).Substring(0, 10)
 $gitBranch    = (git rev-parse --abbrev-ref HEAD)
 $timeStamp    = Get-Date -Format 'yyyyMMddHHmm'
-$versionFull  = "1.0.0-beta1"
+$versionInfo  = "1.0.0-beta1"
 $banner       = ""
 
 # Folders
@@ -36,47 +48,87 @@ $packageFolder = Join-Path -Path $rootPath -ChildPath "dist"
 # Methods
 
 function Prepare-Version {
-    $script:version="1.3.0"
-    $script:versionFull="$version-rc1"
+    $tmp    = ""
+    $source = ""
+    if (!$version) {
+        $tag = (git describe --abbrev=0)
+        if (!$tag) {
+            Write-Host "No Git TAG defined!"
+            exit 1
+        }         
+        $tmp=$tag
+        $source="Git"
+    } else {
+        $tmp=$version
+        $source="parameters"
+    }
+    
+    # Sanitize version, remove extra v
+    if ($tmp.StartsWith("v")) { $tmp=$tmp.SubString(1) }
+    
+    # Check version format
+    $re=[regex]"([0-9]\.[0-9]\.[0-9])(\-){0,1}((.)*)"
+    $m=$re.Match($tmp)
+    if (!$m.Success) {
+        Write-Host "Invalid version format! $tmp"
+        exit 2
+    }
+    
+    $script:version = $m.Groups[1].Value
+    $info = $m.Groups[3].Value
+    if ($info -eq "") {
+        $script:versionInfo="$script:version"
+    } else {
+        
+        $script:versionInfo="$script:version-$info"
+    }
+    
+    Write-Host "Version $script:version, full version $script:versionInfo (from $source)..."
+    Write-Host "Git commit $gitHash, branch $gitBranch"
 }    
 
+# Check for GIT pending changes
+function Pending-Changes {
+    if ($allowDirty) { return }
+    
+    $st = (git status -su)
+    if (![string]::ISNullOrEmpty($st)) {
+        Write-Host "`n`nGit repository has pending changes!`nAborting..."
+        exit 3
+    }
+}
 # Initialize folder if missing
 function Initialize-Folder {
     param ([string] $folderName)
-    if (!(test-path $folderName)) {
+    if (!(Test-Path $folderName)) {
         New-Item -ItemType Directory -Force -Path $folderName | Out-Null
     }
+    Write-Host "Output folder '$outputFolder'`n`n"
 }
 
 # Clear temp folder
 function Clear-Folder {
     param ([string] $path)
-    if (Test-Path $path) {
-        Remove-Item $path -Recurse
-    }
+    if (Test-Path $path) { Remove-Item $path -Recurse }
 }
 
 # Initialize banner variable
 function Initialize-Banner {
     $script:banner = "
-   ______      __                 ______              
-  / ____/_  __/ /_  ___  _____   / ____/___  ________ 
- / /   / / / / __ \/ _ \/ ___/  / /   / __ \/ ___/ _ \
-/ /___/ /_/ / /_/ /  __(__  )  / /___/ /_/ / /  /  __/
-\____/\__,_/_.___/\___/____/   \____/\____/_/   \___/ 
+   _____ ______         _____           _              _____ _____ 
+  / ____|  ____|       |  __ \         | |       /\   |  __ \_   _|
+ | (___ | |__   _ __   | |__) |___  ___| |_     /  \  | |__) || |  
+  \___ \|  __| | '_ \  |  _  // _ \/ __| __|   / /\ \ |  ___/ | |  
+  ____) | |____| | | | | | \ \  __/\__ \ |_   / ____ \| |    _| |_ 
+ |_____/|______|_| |_| |_|  \_\___||___/\__| /_/    \_\_|   |_____|
 
 
-Commit         : $gitHash
-Branch         : $gitBranch "
+Packaging...
+"
 }
 
 # Display banner
-function Display-Banner {
-    
-    Write-Host $banner 
-    Write-Host "Output folder  : $outputFolder"
-    Write-Host "`n`n"
-}
+function Display-Banner { Write-Host $banner }
 
 # Build
 function Build-Solution {
@@ -87,7 +139,7 @@ function Build-Solution {
         -t:Rebuild `
         -p:Configuration=Release `
         -p:Version=$version `
-        -p:versionFull=$versionFull `
+        -p:versionInfo=$versionInfo `
         -p:OutDir=$outputFolder -v:n
 }
 
@@ -102,30 +154,35 @@ function Write-BuildInformation {
 
 # Create ZIP file
 function Create-Package {
-    if (!(test-path $packageFolder)) {
+    if (!(Test-Path $packageFolder)) {
         New-Item -ItemType Directory -Force -Path $packageFolder | Out-Null
     }
     
     Compress-Archive `
         -Path "$outputFolder\*" `
-        -Destination "$packageFolder\$project-v$versionFull.zip" `
+        -Destination "$packageFolder\$project-v$versionInfo.zip" `
         -Force
 }
 
 # Deploy package
 function Deploy-Package {
-
+    if ($deploy) {
+        Write-Host Deploying package...
+    }
 }
 # ------------------------------------------------------------------------------
 
 
 
-# If version not defined, get from Git
-Prepare-Version
-
 # Banner, info
 Initialize-Banner
 Display-Banner
+
+# If version not defined, get from Git
+Prepare-Version
+
+# Check for pending changes
+Pending-Changes
 
 # Prepare new output folder
 Initialize-Folder $outputFolder
